requirements.txt
asyncpg==0.29.0
redis==5.0.1
python-dateutil==2.8.2
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
websockets==12.0
bcrypt==4.1.2
PyJWT==2.8.0